+++ components/dashboard/admin/areas/tabs/AreaAmenitiesFAQs.tsx
@@ -56,7 +56,7 @@ export default function AreaAmenitiesFAQs({ areaId, faqs = [] }: AreaAmenitiesFA
   };
 
   return (
-    <WidgetCard className="p-0">
+    <WidgetCard className="px-0 pt-0 pb">
       <form onSubmit={form.handleSubmit(onSubmit)}>
         <FAQTab
           control={form.control}